+++ docx/oxml/section.py
@@ -1,16 +1,34 @@
 # encoding: utf-8
 
-"""
-Section-related custom element classes.
-"""
+"""Section-related custom element classes"""
 
-from __future__ import absolute_import, print_function
+from __future__ import absolute_import, division, print_function, unicode_literals
 
 from copy import deepcopy
 
-from ..enum.section import WD_ORIENTATION, WD_SECTION_START
-from .simpletypes import ST_SignedTwipsMeasure, ST_TwipsMeasure
-from .xmlchemy import BaseOxmlElement, OptionalAttribute, ZeroOrOne
+from docx.enum.section import WD_HEADER_FOOTER, WD_ORIENTATION, WD_SECTION_START
+from docx.oxml.simpletypes import ST_SignedTwipsMeasure, ST_TwipsMeasure, XsdString
+from docx.oxml.xmlchemy import (
+    BaseOxmlElement,
+    OptionalAttribute,
+    RequiredAttribute,
+    ZeroOrMore,
+    ZeroOrOne,
+)
+
+
+class CT_HdrFtr(BaseOxmlElement):
+    """`w:hdr` and `w:ftr`, the root element for header and footer part respectively"""
+
+    p = ZeroOrMore('w:p', successors=())
+    tbl = ZeroOrMore('w:tbl', successors=())
+
+
+class CT_HdrFtrRef(BaseOxmlElement):
+    """`w:headerReference` and `w:footerReference` elements"""
+
+    type_ = RequiredAttribute('w:type', WD_HEADER_FOOTER)
+    rId = RequiredAttribute('r:id', XsdString)
 
 
 class CT_PageMar(BaseOxmlElement):
@@ -38,25 +56,41 @@ class CT_PageSz(BaseOxmlElement):
 
 
 class CT_SectPr(BaseOxmlElement):
-    """
-    ``<w:sectPr>`` element, the container element for section properties.
-    """
-    __child_sequence__ = (
-        'w:footnotePr', 'w:endnotePr', 'w:type', 'w:pgSz', 'w:pgMar',
-        'w:paperSrc', 'w:pgBorders', 'w:lnNumType', 'w:pgNumType', 'w:cols',
-        'w:formProt', 'w:vAlign', 'w:noEndnote', 'w:titlePg',
-        'w:textDirection', 'w:bidi', 'w:rtlGutter', 'w:docGrid',
-        'w:printerSettings', 'w:sectPrChange',
+    """`w:sectPr` element, the container element for section properties"""
+
+    _tag_seq = (
+        'w:footnotePr', 'w:endnotePr', 'w:type', 'w:pgSz', 'w:pgMar', 'w:paperSrc',
+        'w:pgBorders', 'w:lnNumType', 'w:pgNumType', 'w:cols', 'w:formProt', 'w:vAlign',
+        'w:noEndnote', 'w:titlePg', 'w:textDirection', 'w:bidi', 'w:rtlGutter',
+        'w:docGrid', 'w:printerSettings', 'w:sectPrChange',
     )
-    type = ZeroOrOne('w:type', successors=(
-        __child_sequence__[__child_sequence__.index('w:type')+1:]
-    ))
-    pgSz = ZeroOrOne('w:pgSz', successors=(
-        __child_sequence__[__child_sequence__.index('w:pgSz')+1:]
-    ))
-    pgMar = ZeroOrOne('w:pgMar', successors=(
-        __child_sequence__[__child_sequence__.index('w:pgMar')+1:]
-    ))
+    headerReference = ZeroOrMore("w:headerReference", successors=_tag_seq)
+    footerReference = ZeroOrMore("w:footerReference", successors=_tag_seq)
+    type = ZeroOrOne("w:type", successors=_tag_seq[3:])
+    pgSz = ZeroOrOne("w:pgSz", successors=_tag_seq[4:])
+    pgMar = ZeroOrOne("w:pgMar", successors=_tag_seq[5:])
+    titlePg = ZeroOrOne("w:titlePg", successors=_tag_seq[14:])
+    del _tag_seq
+
+    def add_footerReference(self, type_, rId):
+        """Return newly added CT_HdrFtrRef element of *type_* with *rId*.
+
+        The element tag is `w:footerReference`.
+        """
+        footerReference = self._add_footerReference()
+        footerReference.type_ = type_
+        footerReference.rId = rId
+        return footerReference
+
+    def add_headerReference(self, type_, rId):
+        """Return newly added CT_HdrFtrRef element of *type_* with *rId*.
+
+        The element tag is `w:headerReference`.
+        """
+        headerReference = self._add_headerReference()
+        headerReference.type_ = type_
+        headerReference.rId = rId
+        return headerReference
 
     @property
     def bottom_margin(self):
@@ -101,6 +135,23 @@ class CT_SectPr(BaseOxmlElement):
     def footer(self, value):
         pgMar = self.get_or_add_pgMar()
         pgMar.footer = value
+
+    def get_footerReference(self, type_):
+        """Return footerReference element of *type_* or None if not present."""
+        path = "./w:footerReference[@w:type='%s']" % WD_HEADER_FOOTER.to_xml(type_)
+        footerReferences = self.xpath(path)
+        if not footerReferences:
+            return None
+        return footerReferences[0]
+
+    def get_headerReference(self, type_):
+        """Return headerReference element of *type_* or None if not present."""
+        matching_headerReferences = self.xpath(
+            "./w:headerReference[@w:type='%s']" % WD_HEADER_FOOTER.to_xml(type_)
+        )
+        if len(matching_headerReferences) == 0:
+            return None
+        return matching_headerReferences[0]
 
     @property
     def gutter(self):
@@ -154,23 +205,6 @@ class CT_SectPr(BaseOxmlElement):
         pgMar.left = value
 
     @property
-    def right_margin(self):
-        """
-        The value of the ``w:right`` attribute in the ``<w:pgMar>`` child
-        element, as a |Length| object, or |None| if either the element or the
-        attribute is not present.
-        """
-        pgMar = self.pgMar
-        if pgMar is None:
-            return None
-        return pgMar.right
-
-    @right_margin.setter
-    def right_margin(self, value):
-        pgMar = self.get_or_add_pgMar()
-        pgMar.right = value
-
-    @property
     def orientation(self):
         """
         The member of the ``WD_ORIENTATION`` enumeration corresponding to the
@@ -220,6 +254,44 @@ class CT_SectPr(BaseOxmlElement):
         pgSz.w = value
 
     @property
+    def preceding_sectPr(self):
+        """sectPr immediately preceding this one or None if this is the first."""
+        # ---[1] predicate returns list of zero or one value---
+        preceding_sectPrs = self.xpath("./preceding::w:sectPr[1]")
+        return preceding_sectPrs[0] if len(preceding_sectPrs) > 0 else None
+
+    def remove_footerReference(self, type_):
+        """Return rId of w:footerReference child of *type_* after removing it."""
+        footerReference = self.get_footerReference(type_)
+        rId = footerReference.rId
+        self.remove(footerReference)
+        return rId
+
+    def remove_headerReference(self, type_):
+        """Return rId of w:headerReference child of *type_* after removing it."""
+        headerReference = self.get_headerReference(type_)
+        rId = headerReference.rId
+        self.remove(headerReference)
+        return rId
+
+    @property
+    def right_margin(self):
+        """
+        The value of the ``w:right`` attribute in the ``<w:pgMar>`` child
+        element, as a |Length| object, or |None| if either the element or the
+        attribute is not present.
+        """
+        pgMar = self.pgMar
+        if pgMar is None:
+            return None
+        return pgMar.right
+
+    @right_margin.setter
+    def right_margin(self, value):
+        pgMar = self.get_or_add_pgMar()
+        pgMar.right = value
+
+    @property
     def start_type(self):
         """
         The member of the ``WD_SECTION_START`` enumeration corresponding to
@@ -238,6 +310,21 @@ class CT_SectPr(BaseOxmlElement):
             return
         type = self.get_or_add_type()
         type.val = value
+
+    @property
+    def titlePg_val(self):
+        """Value of `w:titlePg/@val` or |None| if not present"""
+        titlePg = self.titlePg
+        if titlePg is None:
+            return False
+        return titlePg.val
+
+    @titlePg_val.setter
+    def titlePg_val(self, value):
+        if value in [None, False]:
+            self._remove_titlePg()
+        else:
+            self.get_or_add_titlePg().val = value
 
     @property
     def top_margin(self):